System logs field
In Funnel, we've found that there are a lot of logs coming from the system and not the executor (container) that are important: an error message from the object store client, an error message from the system about a disk being full, a docker run command, etc. We've been avoiding adding this to TES and putting it in separate endpoints, or accessing these logs via external systems such as rsyslog or Kibana, but I think it's worth proposing here. It's an optional field, and in the logs message, so not much overhead to implementations.